refactor: move IndieAuthModal component
